+++ src/components/StepCommandBuilder.jsx
@@ -8,7 +8,7 @@ import StepStrictRecipeSourcesConfig from './StepStrictRecipeSourcesConfig';
 import StepOrgServiceConfig from './StepOrgServiceConfig';
 
 const steps = [
-  { label: 'Core variables', component: StepGeneralConfig },
+  // { label: 'Core variables', component: StepGeneralConfig },
   { label: 'SCM Configuration', component: StepSCMConfig },
   { label: '(Optional) Organization Service', component: StepOrgServiceConfig },
   { label: '(Optional) Strict Recipe Sources', component: StepStrictRecipeSourcesConfig },
